+++ example.js
@@ -1,9 +1,9 @@
 var sodium = require('./')
 
-var nonce = new Buffer(sodium.crypto_secretbox_NONCEBYTES)
-var key = new Buffer(sodium.crypto_secretbox_KEYBYTES)
-var message = new Buffer('Hello, World!')
-var cipher = new Buffer(message.length + sodium.crypto_secretbox_MACBYTES)
+var nonce = Buffer.alloc(sodium.crypto_secretbox_NONCEBYTES)
+var key = sodium.sodium_malloc(sodium.crypto_secretbox_KEYBYTES)
+var message = Buffer.from('Hello, World!')
+var cipher = Buffer.alloc(message.length + sodium.crypto_secretbox_MACBYTES)
 
 sodium.randombytes_buf(nonce) // insert random data into nonce
 sodium.randombytes_buf(key)  // insert random data into key
@@ -13,7 +13,7 @@ sodium.crypto_secretbox_easy(cipher, message, nonce, key)
 
 console.log('Encrypted message:', cipher)
 
-var plainText = new Buffer(cipher.length - sodium.crypto_secretbox_MACBYTES)
+var plainText = Buffer.alloc(cipher.length - sodium.crypto_secretbox_MACBYTES)
 
 if (!sodium.crypto_secretbox_open_easy(plainText, cipher, nonce, key)) {
   console.log('Decryption failed!')