+++ test/crypto_box.js
@@ -1,14 +1,10 @@
 var tape = require('tape')
 var sodium = require('../')
-var alloc = require('buffer-alloc')
-var fill = require('buffer-fill')
 
 tape('crypto_box_seed_keypair', function (t) {
-  var pk = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk = alloc(sodium.crypto_box_SECRETKEYBYTES)
-  var seed = alloc(sodium.crypto_box_SEEDBYTES)
-
-  fill(seed, 'lo')
+  var pk = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var seed = Buffer.alloc(sodium.crypto_box_SEEDBYTES, 'lo')
 
   t.throws(function () {
     sodium.crypto_box_seed_keypair()
@@ -29,13 +25,13 @@ tape('crypto_box_seed_keypair', function (t) {
 })
 
 tape('crypto_box_keypair', function (t) {
-  var pk = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk = alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var pk = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
 
   sodium.crypto_box_keypair(pk, sk)
 
-  t.notEqual(pk, alloc(pk.length), 'made public key')
-  t.notEqual(sk, alloc(sk.length), 'made secret key')
+  t.notEqual(pk, Buffer.alloc(pk.length), 'made public key')
+  t.notEqual(sk, Buffer.alloc(sk.length), 'made secret key')
 
   t.throws(function () {
     sodium.crypto_box_keypair()
@@ -49,22 +45,22 @@ tape('crypto_box_keypair', function (t) {
 })
 
 tape('crypto_box_detached', function (t) {
-  var pk = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk = alloc(sodium.crypto_box_SECRETKEYBYTES)
-  var nonce = alloc(sodium.crypto_box_NONCEBYTES)
+  var pk = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var nonce = Buffer.alloc(sodium.crypto_box_NONCEBYTES)
 
   sodium.crypto_box_keypair(pk, sk)
 
   var message = new Buffer('Hello, World!')
-  var mac = alloc(sodium.crypto_box_MACBYTES)
-  var cipher = alloc(message.length)
+  var mac = Buffer.alloc(sodium.crypto_box_MACBYTES)
+  var cipher = Buffer.alloc(message.length)
 
   sodium.crypto_box_detached(cipher, mac, message, nonce, pk, sk)
 
-  t.notEqual(cipher, alloc(cipher.length), 'not blank')
+  t.notEqual(cipher, Buffer.alloc(cipher.length), 'not blank')
 
-  var plain = alloc(cipher.length)
-  t.notOk(sodium.crypto_box_open_detached(plain, cipher, alloc(mac.length), nonce, pk, sk), 'does not decrypt')
+  var plain = Buffer.alloc(cipher.length)
+  t.notOk(sodium.crypto_box_open_detached(plain, cipher, Buffer.alloc(mac.length), nonce, pk, sk), 'does not decrypt')
   t.ok(sodium.crypto_box_open_detached(plain, cipher, mac, nonce, pk, sk), 'decrypts')
   t.same(plain, message, 'same message')
 
@@ -72,21 +68,21 @@ tape('crypto_box_detached', function (t) {
 })
 
 tape('crypto_box_easy', function (t) {
-  var pk = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk = alloc(sodium.crypto_box_SECRETKEYBYTES)
-  var nonce = alloc(sodium.crypto_box_NONCEBYTES)
+  var pk = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var nonce = Buffer.alloc(sodium.crypto_box_NONCEBYTES)
 
   sodium.crypto_box_keypair(pk, sk)
 
   var message = new Buffer('Hello, World!')
-  var cipher = alloc(message.length + sodium.crypto_box_MACBYTES)
+  var cipher = Buffer.alloc(message.length + sodium.crypto_box_MACBYTES)
 
   sodium.crypto_box_easy(cipher, message, nonce, pk, sk)
 
-  t.notEqual(cipher, alloc(cipher.length), 'not blank')
+  t.notEqual(cipher, Buffer.alloc(cipher.length), 'not blank')
 
-  var plain = alloc(cipher.length - sodium.crypto_box_MACBYTES)
-  t.notOk(sodium.crypto_box_open_easy(plain, alloc(cipher.length), nonce, pk, sk), 'does not decrypt')
+  var plain = Buffer.alloc(cipher.length - sodium.crypto_box_MACBYTES)
+  t.notOk(sodium.crypto_box_open_easy(plain, Buffer.alloc(cipher.length), nonce, pk, sk), 'does not decrypt')
   t.ok(sodium.crypto_box_open_easy(plain, cipher, nonce, pk, sk), 'decrypts')
   t.same(plain, message, 'same message')
 
@@ -94,25 +90,25 @@ tape('crypto_box_easy', function (t) {
 })
 
 tape('crypto_box_seal', function (t) {
-  var pk = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk = alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var pk = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
 
   sodium.crypto_box_keypair(pk, sk)
 
-  var pk2 = alloc(sodium.crypto_box_PUBLICKEYBYTES)
-  var sk2 = alloc(sodium.crypto_box_SECRETKEYBYTES)
+  var pk2 = Buffer.alloc(sodium.crypto_box_PUBLICKEYBYTES)
+  var sk2 = Buffer.alloc(sodium.crypto_box_SECRETKEYBYTES)
 
   sodium.crypto_box_keypair(pk2, sk2)
 
   var message = new Buffer('Hello, sealed World!')
-  var cipher = alloc(message.length + sodium.crypto_box_SEALBYTES)
+  var cipher = Buffer.alloc(message.length + sodium.crypto_box_SEALBYTES)
 
   sodium.crypto_box_seal(cipher, message, pk)
   t.notEqual(cipher, message, 'did not encrypt!')
 
-  t.notEqual(cipher, alloc(cipher.length), 'not blank')
+  t.notEqual(cipher, Buffer.alloc(cipher.length), 'not blank')
 
-  var plain = alloc(cipher.length - sodium.crypto_box_SEALBYTES)
+  var plain = Buffer.alloc(cipher.length - sodium.crypto_box_SEALBYTES)
   t.notOk(sodium.crypto_box_seal_open(plain, cipher, pk2, sk2), 'does not decrypt')
   t.ok(sodium.crypto_box_seal_open(plain, cipher, pk, sk), 'decrypts')
   t.same(plain, message, 'same message')