+++ test/crypto_hash_sha256.js
@@ -1,13 +1,12 @@
 var tape = require('tape')
 var sodium = require('../')
-var alloc = require('buffer-alloc')
 
 tape('crypto_hash_sha256', function (t) {
-  var out = alloc(sodium.crypto_hash_sha256_BYTES)
-  var inp = new Buffer('Hej, Verden!')
+  var out = Buffer.alloc(sodium.crypto_hash_sha256_BYTES)
+  var inp = Buffer.from('Hej, Verden!')
 
   t.throws(function () {
-    sodium.crypto_hash(alloc(0), inp)
+    sodium.crypto_hash(Buffer.alloc(0), inp)
   }, 'throws on bad input')
 
   sodium.crypto_hash_sha256(out, inp)
@@ -19,8 +18,8 @@ tape('crypto_hash_sha256', function (t) {
 })
 
 tape('crypto_hash_sha256_instance', function (t) {
-  var out = alloc(sodium.crypto_hash_sha256_BYTES)
-  var inp = new Buffer('Hej, Verden!')
+  var out = Buffer.alloc(sodium.crypto_hash_sha256_BYTES)
+  var inp = Buffer.from('Hej, Verden!')
 
   var instance = sodium.crypto_hash_sha256_instance()
   instance.update(inp)