+++ test/crypto_scalarmult.js
@@ -1,12 +1,11 @@
 var tape = require('tape')
-var alloc = require('buffer-alloc')
 var sodium = require('../')
 
 tape('crypto_scalarmult_base', function (t) {
   var keys = keyPair()
 
-  t.notEqual(keys.secretKey, alloc(keys.secretKey.length), 'secret key not blank')
-  t.notEqual(keys.publicKey, alloc(keys.publicKey.length), 'public key not blank')
+  t.notEqual(keys.secretKey, Buffer.alloc(keys.secretKey.length), 'secret key not blank')
+  t.notEqual(keys.publicKey, Buffer.alloc(keys.publicKey.length), 'public key not blank')
   t.end()
 })
 
@@ -17,8 +16,8 @@ tape('crypto_scalarmult', function (t) {
   t.notEqual(peer1.secretKey, peer2.secretKey, 'diff secret keys')
   t.notEqual(peer1.publicKey, peer2.publicKey, 'diff public keys')
 
-  var shared1 = alloc(sodium.crypto_scalarmult_BYTES)
-  var shared2 = alloc(sodium.crypto_scalarmult_BYTES)
+  var shared1 = Buffer.alloc(sodium.crypto_scalarmult_BYTES)
+  var shared2 = Buffer.alloc(sodium.crypto_scalarmult_BYTES)
 
   sodium.crypto_scalarmult(shared1, peer1.secretKey, peer2.publicKey)
   sodium.crypto_scalarmult(shared2, peer2.secretKey, peer1.publicKey)
@@ -29,10 +28,10 @@ tape('crypto_scalarmult', function (t) {
 })
 
 function keyPair () {
-  var secretKey = alloc(sodium.crypto_scalarmult_SCALARBYTES)
+  var secretKey = Buffer.alloc(sodium.crypto_scalarmult_SCALARBYTES)
   sodium.randombytes_buf(secretKey)
 
-  var publicKey = alloc(sodium.crypto_scalarmult_BYTES)
+  var publicKey = Buffer.alloc(sodium.crypto_scalarmult_BYTES)
   sodium.crypto_scalarmult_base(publicKey, secretKey)
 
   return {