+++ test/randombytes.js
@@ -1,6 +1,5 @@
 var tape = require('tape')
 var sodium = require('../')
-var alloc = require('buffer-alloc')
 
 tape('constants', function (assert) {
   assert.same(typeof sodium.randombytes_SEEDBYTES, 'number', 'randombytes_SEEDBYTES is number')
@@ -30,13 +29,13 @@ tape('randombytes_uniform', function (t) {
 tape('randombytes_buf', function (t) {
   var buf = null
 
-  buf = alloc(10)
+  buf = Buffer.alloc(10)
   sodium.randombytes_buf(buf)
-  t.notEqual(buf, alloc(10), 'not blank')
+  t.notEqual(buf, Buffer.alloc(10), 'not blank')
 
-  buf = alloc(1024)
+  buf = Buffer.alloc(1024)
   sodium.randombytes_buf(buf)
-  t.notEqual(buf, alloc(1024), 'large not blank')
+  t.notEqual(buf, Buffer.alloc(1024), 'large not blank')
 
   t.end()
 })
@@ -44,8 +43,8 @@ tape('randombytes_buf', function (t) {
 tape('randombytes_deterministic', function (t) {
   var seed1 = Buffer.allocUnsafe(sodium.randombytes_SEEDBYTES)
   var seed2 = Buffer.allocUnsafe(sodium.randombytes_SEEDBYTES)
-  var buf1 = alloc(10)
-  var buf2 = alloc(10)
+  var buf1 = Buffer.alloc(10)
+  var buf2 = Buffer.alloc(10)
 
   for (var i = 0; i < 1e6; i++) {
     sodium.randombytes_buf(seed1)